pyproject.toml
[project]
name = "agente-conteudo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.4.3",
    "python-dotenv>=1.2.1",
    "groq>=0.11.0",
    "ffmpeg>=1.4",
    "openai>=2.15.0",
    "tavily-python>=0.7.19",
    "sqlalchemy>=2.0.46",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "jwt>=1.4.0",
    "pyjwt>=2.10.1",
]
